% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSjags1.R
\name{make_SS_data}
\alias{make_SS_data}
\title{Prepare the Data for a State-Space Model}
\usage{
make_SS_data(
  y,
  x = NULL,
  stochasticPeriods = NULL,
  deterministicPeriods = NULL,
  AR1 = FALSE,
  sig_eps_prior = NULL,
  sig_xi_prior = NULL,
  sig_omega_prior = NULL,
  sig_psi_prior = NULL,
  phi_prior = "dunif(0,1)",
  sigeps_breaks = NULL,
  sigxi_breaks = NULL,
  normalizedRate = FALSE
)
}
\arguments{
\item{y}{The input time series, expressed as a numeric vector.}

\item{x}{The corresponding time measurements, expressed as a numeric vector.
If the default \code{NULL} is used, the time measurements will be assumed to be
equally-spaced integer values.}

\item{stochasticPeriods}{An optional vector of stochastic cycle periods, in
units of time.  Defaults to \code{NULL}, indicating no stochastic cycle present.}

\item{deterministicPeriods}{An optional vector of deterministic cycle periods, in
units of time.  Defaults to \code{NULL}, indicating no deterministic cycle present.}

\item{AR1}{Whether to include an AR(1) autoregressive process.
Defaults to \code{FALSE}.}

\item{sig_eps_prior}{The prior used for the IRREGULAR standard deviation(s),
expressed in 'JAGS' syntax.  Note that providing a number here will set this value
to a constant, instead of treating it as a modeled quantity.  If the default \code{NULL}
is used, an exponential prior will be used (\code{dexp()} in 'JAGS' syntax), with
a rate parameter determined by the approximate scale of the data.}

\item{sig_xi_prior}{The prior used for the RATE DISTURBANCE standard deviation(s),
expressed in 'JAGS' syntax.  Note that providing a number here will set this value
to a constant, instead of treating it as a modeled quantity.  If the default \code{NULL}
is used, an exponential prior will be used (\code{dexp()} in 'JAGS' syntax), with
a rate parameter determined by the approximate scale of the data.}

\item{sig_omega_prior}{The prior used for the CYCLIC DISTURBANCE standard deviation(s),
expressed in 'JAGS' syntax.  Note that providing a number here will set this value
to a constant, instead of treating it as a modeled quantity.  Note that this argument
will be ignored if \code{stochasticPeriods} is set to \code{NULL}.   If the default \code{NULL}
is used, an exponential prior will be used (\code{dexp()} in 'JAGS' syntax), with
a rate parameter determined by the approximate scale of the data.}

\item{sig_psi_prior}{The prior used for the AR(1) PROCESS DISTURBANCE standard deviation,
expressed in 'JAGS' syntax.  Note that providing a number here will set this value
to a constant, instead of treating it as a modeled quantity.  Note that this argument
will be ignored unless \code{AR1} is set to \code{process}.  If the default \code{NULL}
is used, an exponential prior will be used (\code{dexp()} in 'JAGS' syntax), with
a rate parameter determined by the approximate scale of the data.}

\item{phi_prior}{The prior used for the AR(1) autoregressive parameter,
expressed in 'JAGS' syntax.  Note that providing a number here will set this value
to a constant, instead of treating it as a modeled quantity.  Note that this argument
will be ignored if \code{AR1==FALSE}.  Defaults to \code{"dunif(0,1)"}}

\item{sigeps_breaks}{An optional vector of structural breakpoints in the irregular
component, which may be interpreted as different irregular standard deviations
in different time periods.  Breakpoints should be expressed in the units of the
time measurements supplied to argument \code{x}.  A single breakpoint may be used if
there are two time periods, etc.  Defaults to \code{NULL}, indicating no breaks.}

\item{sigxi_breaks}{An optional vector of structural breakpoints in the RATE
DISTURBANCE component, which may be interpreted as different RATE
DISTURBANCE standard deviations
in different time periods.  Breakpoints should be expressed in the units of the
time measurements supplied to argument \code{x}.  A single breakpoint may be used if
there are two time periods, etc.  Defaults to \code{NULL}, indicating no breaks.}

\item{normalizedRate}{Whether to express treat the RATE DISTURBANCES as occurring
on the normalized time scale, as opposed to the data time scale.  If this is set
to \code{TRUE}, the rates and rate disturbance standard deviation may be interpreted
as rate per normalized time step; If this is set
to \code{FALSE}, the rates and rate disturbance standard deviation may be interpreted
as rate per unit time with respect to the input data units.  This functionality
is retained in order to facilitate comparison with output from other software.
Defaults to \code{TRUE}.}
}
\value{
A list with the following components:
\itemize{
\item \code{y}: The input time series, expressed as a numeric vector
\item \code{n}: The length of the input time series
\item \code{startpoint}: The starting value of the time series
\item \code{trendprecinit}: The prior precision used for the initial value of trend
\item \code{rateprecinit}: The prior precision used for the initial value of rate
\item \code{sigeps_split}: A vector corresponding to the index of irregular standard deviation
parameter used for each time series element
\item \code{n_sigeps_split}: The number of irregular standard devation parameters to estimate
\item \code{sigxi_split}: A vector corresponding to the index of rate disturbance standard deviation
parameter used for each time series element
\item \code{n_sigxi_split}: The number of rate disturbance standard devation parameters to estimate
\item \code{phi_prior}: The prior used for autoregressive parameter phi
\item \code{p_s}: Stochastic periods, expressed as a numeric vector
\item \code{n_ps}: Number of stochastic periods
\item \code{p_d}: Deterministic periods, expressed as a numeric vector
\item \code{n_pd}: Number of deterministic periods
\item \code{sig_eps_prior_num}: Irregular standard deviation, if a constant (or vector of constants) is used
\item \code{sig_xi_prior_num}: Rate disturbance standard deviation, if a constant (or vector of constants) is used
\item \code{sig_eps_prior_num}: Cycle standard deviation, if a constant (or vector of constants) is used
}
}
\description{
Prepares a data object to be used by \link{runSS}, also called
internally within \link{runSS}.  The principal utility of

The user may choose to use the model created by \link{runSS} as a starting
point and make adjustments manually, in which case having the processed data
at hand may make such adjustments more straightforward.  Additionally, it may
be useful to have a check that the appropriate data is being used by the model.
}
\examples{
make_SS_data(y = SS_data$y,                   # vector of observations
             x = SS_data$x,                   # associated time measurements
             stochasticPeriods = c(1, 0.5)),  # stochastic cycle periods
}
\author{
Matt Tyers
}
