% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSjags1.R
\name{psd_components}
\alias{psd_components}
\title{Lomb-Scargle Power Spectral Density of State-space Model Components}
\usage{
psd_components(
  jagsout,
  y = NULL,
  x = NULL,
  alpha = 0.01,
  collapsecycle = FALSE,
  ...
)
}
\arguments{
\item{jagsout}{Output object returned from \link{runSS}.  Note that this will
be an object of class \code{jagsUI}.}

\item{y}{Input time series used by \link{runSS}.  Used by function to calculate
the Irregular component.  Accepting the default value \code{NULL} will omit plotting
the Irregular component.}

\item{x}{Time measurements associated with time series \code{y}.  If
default value \code{NULL} is accepted, integer-valued time steps will be plotted.}

\item{alpha}{Significance threshold for reference lines.  Defaults to \code{0.01}.}

\item{collapsecycle}{Whether to collapse all stochastic and/or deterministic
cycle components as one single Cycle component}

\item{...}{additional plotting arguments}
}
\description{
Produces a plot of the respective components of a State-space model,
according to the Lomb-Scargle Power Spectral Density (PSD).

Possible components include Trend, Cycle (stochastic and/or deterministic), AR(1),
and Irregular.
}
\examples{
psd_components(jagsout=SS_out, y=SS_data$y, x=SS_data$x)
psd_components(jagsout=SS_out, y=SS_data$y, x=SS_data$x, collapsecycle=TRUE)
}
\author{
Matt Tyers
}
