% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSjags1.R
\docType{data}
\name{SS_data}
\alias{SS_data}
\title{Example State-space input data}
\format{
An object of class \code{data.frame} with 49 rows and 2 columns.
}
\usage{
SS_data
}
\description{
DOCUMENT ME!
}
\keyword{datasets}
