\name{SSjags-package}
\alias{SSjags-package}
\alias{SSjags}
\docType{package}
\title{
Construct and Fit a State-Space Model in 'JAGS' through 'jagsUI'
}
\description{
The main functionality is a wrapper function (\link{runSS}) that uses user input
    to write model code in JAGS syntax to a temporary text file, and then
    optionally run that model with a user-supplied vector of time-series data
    and optionally corresponding time measurements.  A 'jagsUI' output object
    is then returned.  Functions are also provided for visualization of the
    respective model components.
}
\details{
\tabular{ll}{
Package: \tab SSjags\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2022-11-02\cr
License: \tab GPL-2\cr
}
}
\author{
Matt Tyers

Maintainer: Matt Tyers <matttyersstat@gmail.com>
}
